requirements.txt
Flask
torch
torchvision
numpy
pillow
flask-cors